+++ lib/faro.dart
@@ -12,6 +12,7 @@ export './src/models/models.dart';
 export './src/offline_transport/offline_transport.dart';
 export './src/session/sampling_context.dart';
 export './src/tracing/span.dart';
+export './src/tracing/span_exception_options.dart';
 export './src/transport/faro_transport.dart';
 export './src/user_actions/constants.dart';
 export './src/user_actions/start_user_action_options.dart';